Added toastr animation and popup for confirmation of actions

+++ src/app/courses/courses-list/courses-list.component.ts
@@ -1,13 +1,18 @@
 import { Component, OnInit } from "@angular/core";
 import { CourseService } from "../../shared/course.service";
 import { Course } from "src/app/shared/course.model";
+import { ToastrService } from "ngx-toastr";
+
 @Component({
   selector: "app-courses-list",
   templateUrl: "./courses-list.component.html",
   styleUrls: ["./courses-list.component.css"],
 })
 export class CoursesListComponent implements OnInit {
-  constructor(public courseservice: CourseService) {}
+  constructor(
+    public courseservice: CourseService,
+    private toastr: ToastrService
+  ) {}
 
   ngOnInit(): void {
     this.courseservice.refreshList();
@@ -19,8 +24,11 @@ export class CoursesListComponent implements OnInit {
   }
 
   deleteRecord(id: number) {
-    this.courseservice.deleteCourse(id).subscribe((res) => {
-      this.courseservice.refreshList();
-    });
+    if (confirm("Are you sure you want to delete the record ")) {
+      this.courseservice.deleteCourse(id).subscribe((res) => {
+        this.toastr.warning("Deleted Successfully", "Course Updation");
+        this.courseservice.refreshList();
+      });
+    }
   }
 }
